fixed serial port reading being out of sync (erroneous GPS coordinates)
The GPS signal was giving us delayed data. This was a problem with the serial port data and our time delay function. The time delay function waited every 25 seconds, while new data in the serial port was streaming every second (while we were reading every 25 seconds). We had to find a way to go to the “eof,” or continuously read the serial while executing the code that send data to the server every 25 seconds. Hence, we execute only a specific block of code every 25 seconds and, allowing us to read from the serial port every second.